connection: Refactor connection to use the net-sftp gem instead of net-scp
This allows the connection module to be used for SFTP servers instead of full SSH servers, which reduces the attack surface.

The `net-scp` gem is replaced by the `net-sftp` gem. Also, version constraints are added to the gem dependencies.

Any retry and fallback mechanisms are removed, here, as errors should be handled on a higher level for better robustness.

Signed-off-by: David Stritzl <d.stritzl@dynalynx.nl>