connection: Implement file organisation optimisation
When fetching files from the remote, only files modified after (or equal to) a given time are retrieved. Due to moving from SCP/SSH to SFTP, the necessary file stats for this have to be performed on the client-side. For directories with a large number of files this can have a significant impact on performance.

Here, a file organisation mechanism is implemented to improve the performance of listing files based on the modification time. This works by moving all new files into a YEAR/MONTH/DAY/HOUR directory structure. When listing files based on the modification date, `stat` only has to be done for files in the HOUR directory matching the given date. For all directories past this data, a simple directory list is sufficient.

When fetching a large amount of past data, this can result in a lot of directory lists, however, for the intended use this not expected to be the case.

Signed-off-by: David Stritzl <d.stritzl@dynalynx.nl>